Change types for PTE_{}

+++ kernel-rs/src/proc.rs
@@ -330,7 +330,7 @@ pub unsafe fn procinit() {
             panic(b"kalloc\x00" as *const u8 as *mut u8);
         }
         let va: usize = kstack(p.offset_from(PROC.as_mut_ptr()) as i64 as i32) as usize;
-        kvmmap(va, pa as usize, PGSIZE, (PTE_R | PTE_W) as i32);
+        kvmmap(va, pa as usize, PGSIZE, PTE_R | PTE_W);
         (*p).kstack = va;
         p = p.offset(1)
     }
@@ -452,7 +452,7 @@ pub unsafe fn proc_pagetable(p: *mut Proc) -> PagetableT {
         TRAMPOLINE,
         PGSIZE,
         trampoline.as_mut_ptr() as usize,
-        (PTE_R | PTE_X) as i32,
+        PTE_R | PTE_X,
     );
 
     // map the trapframe just below TRAMPOLINE, for trampoline.S.
@@ -461,7 +461,7 @@ pub unsafe fn proc_pagetable(p: *mut Proc) -> PagetableT {
         TRAPFRAME,
         PGSIZE,
         (*p).tf as usize,
-        (PTE_R | PTE_W) as i32,
+        PTE_R | PTE_W,
     );
     pagetable
 }
